requirements.txt
face_recognition==1.3.0
scipy
Flask==2.0.2
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.0.0
gunicorn

